+++ examples/data_factory/copy_blobs/test_data_factory_copy_blobs_unit.py
@@ -1,8 +1,7 @@
 import pytest
-from data_factory_testing_framework._test_framework import TestFramework, TestFrameworkType
-from data_factory_testing_framework.models.activities.activity import Activity
-from data_factory_testing_framework.models.activities.for_each_activity import ForEachActivity
-from data_factory_testing_framework.models.pipeline import Pipeline
+from data_factory_testing_framework import TestFramework, TestFrameworkType
+from data_factory_testing_framework.models._pipeline import Pipeline
+from data_factory_testing_framework.models.activities import Activity, ForEachActivity
 from data_factory_testing_framework.state import (
     PipelineRunState,
     PipelineRunVariable,
