+++ examples/fabric/batch_job/test_fabric_batchjob_functional.py
@@ -1,5 +1,5 @@
 import pytest
-from data_factory_testing_framework._test_framework import TestFramework, TestFrameworkType
+from data_factory_testing_framework import TestFramework, TestFrameworkType
 from data_factory_testing_framework.state import RunParameterType
 from data_factory_testing_framework.state.run_parameter import RunParameter
