+++ examples/fabric/batch_job/test_fabric_batchjob_unit.py
@@ -1,7 +1,7 @@
 # flake8: noqa: E501
 import pytest
-from data_factory_testing_framework._test_framework import TestFramework, TestFrameworkType
-from data_factory_testing_framework.models.pipeline import Pipeline
+from data_factory_testing_framework import TestFramework, TestFrameworkType
+from data_factory_testing_framework.models._pipeline import Pipeline
 from data_factory_testing_framework.state import (
     PipelineRunState,
     PipelineRunVariable,
