+++ src/data_factory_testing_framework/__init__.py
@@ -1,8 +1,5 @@
 from data_factory_testing_framework._functions.functions_repository import FunctionsRepository
+from data_factory_testing_framework._repositories.data_factory_repository import DataFactoryRepository
 from data_factory_testing_framework._test_framework import TestFramework, TestFrameworkType
 
-__all__ = [
-    "TestFramework",
-    "TestFrameworkType",
-    "FunctionsRepository",
-]
+__all__ = ["TestFramework", "TestFrameworkType", "FunctionsRepository", "DataFactoryRepository"]
