+++ src/data_factory_testing_framework/_deserializers/shared/_activity_deserializer.py
@@ -1,15 +1,17 @@
 from typing import List
 
-from data_factory_testing_framework.models.activities.activity import Activity
-from data_factory_testing_framework.models.activities.append_variable_activity import AppendVariableActivity
-from data_factory_testing_framework.models.activities.execute_pipeline_activity import ExecutePipelineActivity
-from data_factory_testing_framework.models.activities.fail_activity import FailActivity
-from data_factory_testing_framework.models.activities.filter_activity import FilterActivity
-from data_factory_testing_framework.models.activities.for_each_activity import ForEachActivity
-from data_factory_testing_framework.models.activities.if_condition_activity import IfConditionActivity
-from data_factory_testing_framework.models.activities.set_variable_activity import SetVariableActivity
-from data_factory_testing_framework.models.activities.switch_activity import SwitchActivity
-from data_factory_testing_framework.models.activities.until_activity import UntilActivity
+from data_factory_testing_framework.models.activities import (
+    Activity,
+    AppendVariableActivity,
+    ExecutePipelineActivity,
+    FailActivity,
+    FilterActivity,
+    ForEachActivity,
+    IfConditionActivity,
+    SetVariableActivity,
+    SwitchActivity,
+    UntilActivity,
+)
 
 
 def _get_activity_from_activity_data(activity_data: dict) -> Activity:
