+++ src/data_factory_testing_framework/_repositories/_factories/base_repository_factory.py
@@ -1,7 +1,7 @@
 from abc import ABC, abstractmethod
 
-from data_factory_testing_framework.models.pipeline import Pipeline
-from data_factory_testing_framework.repositories.data_factory_repository import DataFactoryRepository
+from data_factory_testing_framework._repositories.data_factory_repository import DataFactoryRepository
+from data_factory_testing_framework.models._pipeline import Pipeline
 
 
 class BaseRepositoryFactory(ABC):
