+++ src/data_factory_testing_framework/_repositories/_factories/fabric_repository_factory.py
@@ -4,8 +4,8 @@ from typing import List
 from data_factory_testing_framework._deserializers._deserializer_fabric import (
     parse_fabric_pipeline_from_pipeline_json_files,
 )
-from data_factory_testing_framework.models.pipeline import Pipeline
-from data_factory_testing_framework.repositories.base_repository_factory import BaseRepositoryFactory
+from data_factory_testing_framework._repositories._factories.base_repository_factory import BaseRepositoryFactory
+from data_factory_testing_framework.models._pipeline import Pipeline
 
 
 class FabricRepositoryFactory(BaseRepositoryFactory):
