+++ src/data_factory_testing_framework/_repositories/data_factory_repository.py
@@ -1,7 +1,7 @@
 from typing import List
 
 from data_factory_testing_framework.exceptions.pipeline_not_found_error import PipelineNotFoundError
-from data_factory_testing_framework.models.pipeline import Pipeline
+from data_factory_testing_framework.models._pipeline import Pipeline
 
 
 class DataFactoryRepository:
