+++ src/data_factory_testing_framework/models/__init__.py
@@ -1,31 +1,4 @@
-from data_factory_testing_framework.models.activities.activity import Activity
-from data_factory_testing_framework.models.activities.activity_dependency import ActivityDependency
-from data_factory_testing_framework.models.activities.append_variable_activity import AppendVariableActivity
-from data_factory_testing_framework.models.activities.control_activity import ControlActivity
-from data_factory_testing_framework.models.activities.execute_pipeline_activity import ExecutePipelineActivity
-from data_factory_testing_framework.models.activities.fail_activity import FailActivity
-from data_factory_testing_framework.models.activities.filter_activity import FilterActivity
-from data_factory_testing_framework.models.activities.for_each_activity import ForEachActivity
-from data_factory_testing_framework.models.activities.if_condition_activity import IfConditionActivity
-from data_factory_testing_framework.models.activities.set_variable_activity import SetVariableActivity
-from data_factory_testing_framework.models.activities.switch_activity import SwitchActivity
-from data_factory_testing_framework.models.activities.until_activity import UntilActivity
-from data_factory_testing_framework.models.data_factory_element import DataFactoryElement
-from data_factory_testing_framework.models.pipeline import Pipeline
+from data_factory_testing_framework.models._data_factory_element import DataFactoryElement
+from data_factory_testing_framework.models._pipeline import Pipeline
 
-__all__ = [
-    "Activity",
-    "ActivityDependency",
-    "AppendVariableActivity",
-    "ControlActivity",
-    "ExecutePipelineActivity",
-    "FailActivity",
-    "FilterActivity",
-    "ForEachActivity",
-    "IfConditionActivity",
-    "SetVariableActivity",
-    "SwitchActivity",
-    "UntilActivity",
-    "DataFactoryElement",
-    "Pipeline",
-]
+__all__ = ["DataFactoryElement", "Pipeline"]
