+++ src/data_factory_testing_framework/models/activities/_execute_pipeline_activity.py
@@ -1,9 +1,9 @@
 from typing import Any, Callable, Iterator, List
 
-from data_factory_testing_framework.models.activities.activity import Activity
-from data_factory_testing_framework.models.activities.control_activity import ControlActivity
-from data_factory_testing_framework.models.data_factory_element import DataFactoryElement
-from data_factory_testing_framework.models.pipeline import Pipeline
+from data_factory_testing_framework.models._data_factory_element import DataFactoryElement
+from data_factory_testing_framework.models._pipeline import Pipeline
+from data_factory_testing_framework.models.activities._activity import Activity
+from data_factory_testing_framework.models.activities._control_activity import ControlActivity
 from data_factory_testing_framework.state import PipelineRunState, RunParameterType
 from data_factory_testing_framework.state.run_parameter import RunParameter
