+++ tests/functional/activity_outputs/test_set_activity_output.py
@@ -1,5 +1,5 @@
 import pytest
-from data_factory_testing_framework._test_framework import TestFramework, TestFrameworkType
+from data_factory_testing_framework import TestFramework, TestFrameworkType
 from data_factory_testing_framework.state.dependency_condition import DependencyCondition
