+++ tests/functional/datafactory_element/test_evaluate_datafactory_element.py
@@ -4,7 +4,7 @@ from data_factory_testing_framework.exceptions.data_factory_element_evaluation_e
 )
 from data_factory_testing_framework.exceptions.parameter_not_found_error import ParameterNotFoundError
 from data_factory_testing_framework.exceptions.user_error import UserError
-from data_factory_testing_framework.models.data_factory_element import DataFactoryElement
+from data_factory_testing_framework.models._data_factory_element import DataFactoryElement
 from data_factory_testing_framework.state.pipeline_run_state import PipelineRunState
