+++ tests/unit/models/activities/control_activities/test_switch_activity.py
@@ -1,9 +1,8 @@
 import pytest
-from data_factory_testing_framework._test_framework import TestFramework, TestFrameworkType
-from data_factory_testing_framework.models.activities.set_variable_activity import SetVariableActivity
-from data_factory_testing_framework.models.activities.switch_activity import SwitchActivity
-from data_factory_testing_framework.models.data_factory_element import DataFactoryElement
-from data_factory_testing_framework.models.pipeline import Pipeline
+from data_factory_testing_framework import TestFramework, TestFrameworkType
+from data_factory_testing_framework.models._data_factory_element import DataFactoryElement
+from data_factory_testing_framework.models._pipeline import Pipeline
+from data_factory_testing_framework.models.activities import SetVariableActivity, SwitchActivity
 from data_factory_testing_framework.state import PipelineRunState
